+++ examples/hello_world.py
@@ -1,5 +1,5 @@
 from jackie.router import Router
-from jackie.http import TextResponse
+from jackie.http import Response
 
 
 app = Router()
@@ -8,4 +8,4 @@ app = Router()
 @app.get('/')
 async def hello_world(request):
     name = request.query.get('name', 'World')
-    return TextResponse(f'Hello, {name}!')
+    return Response(text=f'Hello, {name}!')