+++ jackie/http/request.py
@@ -6,19 +6,58 @@ from ..parse import parse_cookies
 from .stream import Stream
 
 
-# Base class
+def form_body(body):
+    boundary = multipart.generate_boundary()
+    body = multipart.serialize(body, boundary)
+    return body, f'multipart/form-data; boundary={boundary}'
+
+
+def json_body(body):
+    body = json.dumps(body).encode()
+    return body, 'application/json; charset=UTF-8'
+
+
+def text_body(body):
+    body = body.encode()
+    return body, 'text/plain; charset=UTF-8'
+
+
+def base_body(body):
+    return body, None
+
+
+BODY_TYPES = {
+    'form': form_body,
+    'json': json_body,
+    'text': text_body,
+    'body': base_body,
+}
+
 
 class Request(Stream):
 
     def __init__(
-        self, path='/', *, method='GET', query=[], headers=[], body=b'',
-        **kwargs,
+        self, path='/', *, method='GET', query=[], headers=[], **kwargs,
     ):
+        body = None
+        for key, get_body in BODY_TYPES.items():
+            try:
+                data = kwargs.pop(key)
+            except KeyError:
+                continue
+            if body is None:
+                body = get_body(data)
+            else:
+                raise ValueError('multiple body types supplied')
+        body, content_type = body or (b'', None)
+
         super().__init__(body)
         self.path = path
         self.method = method
         self.query = MultiDict(query)
         self.headers = Headers(headers, **kwargs)
+        if content_type is not None:
+            self.headers.setdefault('Content-Type', content_type)
 
         self.router = None
         self.view_name = None
@@ -29,33 +68,3 @@ class Request(Stream):
     @property
     def cookies(self):
         return parse_cookies(self.headers.get('Cookie', ''))
-
-
-# Subclasses
-
-class FormRequest(Request):
-
-    def __init__(self, path='/', body={}, boundary=None, **kwargs):
-        if boundary is None:
-            boundary = multipart.generate_boundary()
-        body = multipart.serialize(body, boundary)
-        super().__init__(path=path, body=body, **kwargs)
-        self.headers.setdefault('Content-Type', (
-            f'multipart/form-data; boundary={boundary}'
-        ))
-
-
-class JsonRequest(Request):
-
-    def __init__(self, path='/', body={}, **kwargs):
-        super().__init__(path=path, body=json.dumps(body).encode(), **kwargs)
-        self.headers.setdefault('Content-Type', (
-            'application/json; charset=UTF-8'
-        ))
-
-
-class TextRequest(Request):
-
-    def __init__(self, path='/', body='', **kwargs):
-        super().__init__(path=path, body=body.encode(), **kwargs)
-        self.headers.setdefault('Content-Type', 'text/plain; charset=UTF-8')